+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   * BraintreeCard
     * Update `BTCard` to make all properties accessible on the initializer only vs via the dot syntax.
     * Remove `BTCardRequest`, use `BTCard` directly instead
+    * Update initializer to `BTCardClient(authorization:)`
   * BraintreePayPal
     * Update `BTPayPalRequest`, `BTPayPalVaultRequest` and `BTPayPalCheckoutRequest` to make all properties accessible on the initializer only vs via the dot syntax.
     * Update PayPal app URL query scheme from `paypal-app-switch-checkout` to `paypal`
@@ -33,8 +34,6 @@
     * Update initializer to `BTAmericanExpressClient(authorization:)`
   * BraintreeApplePay
     * Update initializer to `BTApplePayClient(authorization:)`
-  * BraintreeCard
-    * Update initializer to `BTCardClient(authorization:)`
 
 ## 6.27.0 (2025-01-23)
 * BraintreePayPal